refactor: move spaces zri and uri constructors to spaces api.
[ORQSDK-879]
