+++ src/orquestra/sdk/_base/_spaces/_api.py
@@ -55,3 +55,24 @@ def list_projects(
     runtime = resolved_config._get_runtime()
 
     return runtime.list_projects(resolved_workspace_id)
+
+
+def make_workspace_zri(workspace_id: str) -> str:
+    """
+    Make the workspace ZRI for the specified workspace ID.
+
+    Builds project ZRI from some hardcoded values and the workspaceId based on
+    https://zapatacomputing.atlassian.net/wiki/spaces/Platform/pages/512787664/2022-09-26+Zapata+Resource+Identifiers+ZRIs
+    """  # noqa: E501
+    default_tenant_id = 0
+    special_workspace = "system"
+    zri_type = "resource_group"
+
+    return f"zri:v1::{default_tenant_id}:{special_workspace}:{zri_type}:{workspace_id}"
+
+
+def make_workspace_url(resource_catalog_url: str, workspace_zri: str) -> str:
+    """
+    Construct the URL for a workspace based on the resource catalog and workspace ZRI.
+    """
+    return f"{resource_catalog_url}/api/workspaces/{workspace_zri}"
